即使聊天

+++ Complay/Complay/Utils/LPBadgeView/LPBadgeView.h
@@ -16,9 +16,9 @@
 @interface LPBadgeView : UIView
 
 ///值
-@property (assign, nonatomic) NSInteger value;
+@property (assign, nonatomic) int value;
 ///结束回调
-@property (strong, nonatomic) void (^hiddenBlock)(NSInteger value);
+@property (strong, nonatomic) void (^hiddenBlock)(int value);
 
 ///便利初始化，默认color(主题颜色)为红色
 + (LPBadgeView *)badgeWithColor:(UIColor *)color;
